SourceCode/requirements.txt
absl-py==0.8.0
adjustText==0.7.3
alembic==1.0.11
appdirs==1.4.4
appnope==0.1.0
astor==0.8.0
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.2.0
beautifulsoup4==4.8.0
bioinfokit==0.3
bleach==3.1.0
bokeh==1.4.0
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.0.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
chart-studio==1.0.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.3.0
conda==4.9.0
conda-package-handling==1.7.2
configparser==3.8.1
cryptography==3.1.1
cycler==0.10.0
databricks-cli==0.8.7
ddt==1.2.1
decorator==4.4.0
defusedxml==0.6.0
descartes==1.1.0
Distance==0.1.3
distlib==0.3.1
docker==4.0.2
dynetx==0.2.2
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
findspark==1.3.0
Fiona==1.8.13
Flask==1.1.1
future==0.18.2
fuzzywuzzy==0.17.0
gast==0.3.3
geopandas==0.6.2
gitdb==0.6.4
gitdb2==2.0.5
GitPython==3.0.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googletrans==2.4.0
gorilla==0.3.0
graphviz==0.13
grpcio==1.23.0
gunicorn==19.9.0
gym==0.17.2
h5py==2.10.0
httplib2==0.13.1
idna==2.10
imbalanced-learn==0.5.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.2.3
jupyterlab-server==1.0.6
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
koalas==0.25.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.3.1
matplotlib-venn==0.11.6
mca==1.0.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mlflow==1.2.0
mlxtend==0.17.0
more-itertools==7.2.0
munch==2.5.0
nbconvert==5.6.0
nbformat==4.4.0
netdispatch==0.0.3
networkx==2.4
nltk==3.4.5
notebook==6.0.0
numpy==1.19.1
oauth2==1.9.0.post1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==19.2
pandas==1.1.1
pandocfilters==1.4.2
paramiko==2.7.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.0.0
plotly==4.1.0
pluggy==0.13.0
progressbar2==3.47.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.13.0
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.0
py4j==0.10.7
pyarrow==0.14.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pydotplus==2.0.2
pygam==0.8.0
pyglet==1.5.0
Pygments==2.4.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==2.4.2.post1
pyreadr==0.3.5
pyrsistent==0.15.4
PySocks==1.7.1
pyspark==2.4.4
pytest==5.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python-utils==2.3.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.2
querystring-parser==1.2.4
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
ruamel-yaml==0.15.87
scikit-learn==0.21.3
scipy==1.5.2
scp==0.13.2
seaborn==0.11.0
Send2Trash==1.5.0
Shapely==1.7.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
smmap==0.9.0
smmap2==2.0.5
soupsieve==1.9.4
SQLAlchemy==1.3.7
sqlparse==0.3.0
statsmodels==0.10.1
stepwise-regression==1.0.3
tabulate==0.8.3
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textwrap3==0.9.2
tornado==6.0.4
tqdm==4.50.2
traitlets==4.3.2
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
yellowbrick==1.0.1
yelp==1.0.2
zipp==0.6.0
